シェーダリファクタリング

+++ KIRenderer/DefaultVertexBuffer.cpp
@@ -21,7 +21,7 @@ void DefaultVertexBuffer::SetPosition(GLuint primitiveType, const vector<vec3>& 
 	}
 
 	m_VertexInfo[VERTEX_ATTRIB_POSITION]->Set(position);
-	SetVertexSize(position.size());
+	SetVertexSize((int)position.size());
 	SetPrimitiveType(primitiveType);
 }
